% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_batch_coexp}
\alias{cpp_batch_coexp}
\title{Calculate co-expression for multiple gene pairs}
\usage{
cpp_batch_coexp(st_data_mat, src_genes, dest_genes, cell_indices)
}
\arguments{
\item{st_data_mat}{Expression matrix}

\item{src_genes}{Source gene indices (1-based)}

\item{dest_genes}{Destination gene indices (1-based)}

\item{cell_indices}{Cell indices to use (1-based)}
}
\value{
Co-expression ratios
}
\description{
Calculate co-expression for multiple gene pairs
}
