% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_batch_cor}
\alias{cpp_batch_cor}
\title{Calculate correlation between a vector and matrix columns}
\usage{
cpp_batch_cor(vec1, mat)
}
\arguments{
\item{vec1}{Reference vector}

\item{mat}{Matrix where each column is compared to vec1}
}
\value{
Pearson correlation coefficients
}
\description{
Calculate correlation between a vector and matrix columns
}
