% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_random_walk}
\alias{cpp_random_walk}
\title{Random walk on gene-gene interaction network}
\usage{
cpp_random_walk(
  ggi_src,
  ggi_dest,
  receptor_name,
  tf_names,
  n_walks = 10000L,
  max_hop = 10L,
  seed = 123L
)
}
\arguments{
\item{ggi_src}{Source gene names}

\item{ggi_dest}{Destination gene names}

\item{receptor_name}{Starting receptor name}

\item{tf_names}{TF names to score}

\item{n_walks}{Number of random walks}

\item{max_hop}{Maximum number of hops per walk}

\item{seed}{Random seed for reproducibility}
}
\value{
TF visit frequency scores
}
\description{
Random walk on gene-gene interaction network
}
