% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{createSpaTalk}
\alias{createSpaTalk}
\title{SpaTalk object}
\usage{
createSpaTalk(
  st_data,
  st_meta,
  species,
  if_st_is_sc,
  spot_max_cell,
  celltype = NULL
)
}
\arguments{
\item{st_data}{A data.frame or matrix or dgCMatrix containing counts of spatial transcriptomics, each column representing a spot or a cell, each row representing a gene.}

\item{st_meta}{A data.frame containing coordinate of spatial transcriptomics with three columns, namely \code{'spot'}, \code{'x'}, \code{'y'} for spot-based spatial transcriptomics data or \code{'cell'}, \code{'x'}, \code{'y'} for single-cell spatial transcriptomics data.}

\item{species}{A character meaning species of the spatial transcriptomics data.\code{'Human'} or \code{'Mouse'}.}

\item{if_st_is_sc}{A logical meaning if it is single-cell spatial transcriptomics data. \code{TRUE} is \code{FALSE}.}

\item{spot_max_cell}{A integer meaning max cell number for each plot to predict. If \code{if_st_sc} is \code{FALSE}, please determine the \code{spot_max_cell}. For 10X (55um), we recommend 30. For Slide-seq, we recommend 1.}

\item{celltype}{A character containing the cell type of ST data. To skip the deconvolution step and directly infer cell-cell communication, please define the cell type. Default is \code{NULL}.}
}
\value{
SpaTalk object
}
\description{
create SpaTalk object using spatial transcriptomics data.
}
