% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dec_celltype,SpaTalk-method}
\alias{dec_celltype,SpaTalk-method}
\title{Decomposing cell type for spatial transcriptomics data}
\usage{
\S4method{dec_celltype}{SpaTalk}(
  object,
  sc_data,
  sc_celltype,
  min_percent = 0.5,
  min_nFeatures = 10,
  if_use_normalize_data = T,
  if_use_hvg = F,
  if_use_all_cores = T,
  iter_num = 1000
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{createSpaTalk}}.}

\item{sc_data}{A matrix containing counts of single-cell RNA-seq data as the reference, each column representing a cell, each row representing a gene.}

\item{sc_celltype}{A character containing the cell type of the reference single-cell RNA-seq data.}

\item{min_percent}{Min percent to predict new cell type for single-cell st_data or predict new cell for spot-based st_data. Default is \code{0.5}.}

\item{min_nFeatures}{Min number of expressed features/genes for each spot/cell in \code{st_data}. Default is \code{10}.}

\item{if_use_normalize_data}{Whether to use normalized \code{st_data} and \code{sc_data} with Seurat normalization. Default is \code{TRUE}. set it \code{FALSE} when the st_data and sc_data are already normalized matrix with other methods.}

\item{if_use_hvg}{Whether to use highly variable genes for non-negative regression. Default is \code{FALSE}.}

\item{if_use_all_cores}{Whether to use all CPU cores. Default is \code{TRUE}.}

\item{iter_num}{Number of iteration to genenrate the single-cell data for spot-based data. Default is \code{1000}.}
}
\value{
SpaTalk object containing the decomposing results.
}
\description{
Identify the cellular composition for single-cell or spot-based spatial transcriptomics data with non-negative regression.
}
