% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_dec_result}
\alias{demo_dec_result}
\title{Demo data of dec_result}
\usage{
demo_dec_result()
}
\value{
A matrix.
}
\description{
Demo data of dec_result
}
\details{
\code{dec_result} used in \code{\link{dec_celltype}} must be a \code{matrix} object, each row representing a spot, each column representing a cell type.
}
\examples{
dec_result_demo <- demo_dec_result()
}
