% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_lrpairs}
\alias{demo_lrpairs}
\title{Demo data of lrpairs}
\usage{
demo_lrpairs()
}
\value{
A data.frame.
}
\description{
Demo data of lrpairs
}
\details{
\code{lrpairs} used in \code{\link{dec_cci}} must be a \code{data.frame} object with three columns, namely \code{'ligand'}, \code{'receptor'}, \code{'species'}.
}
\examples{
lrpairs_demo <- demo_lrpairs()
}
