% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_pathways}
\alias{demo_pathways}
\title{Demo data of pathways}
\usage{
demo_pathways()
}
\value{
A data.frame.
}
\description{
Demo data of pathways
}
\details{
\code{pathways} used in \code{\link{dec_cci}} must be a \code{data.frame} object with seven columns, namely \code{'src'}, \code{'dest'}, \code{'pathway'}, \code{'source'}, \code{'type'}, \code{'src_tf'}, \code{'dest_tf'}, \code{'species'}.
}
\examples{
pathways_demo <- demo_pathways()
}
