% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_sc_data}
\alias{demo_sc_data}
\title{Demo data of sc_data}
\usage{
demo_sc_data()
}
\value{
A matrix.
}
\description{
Demo data of sc_data.
}
\details{
\code{sc_data} used in \code{\link{dec_celltype}} must be a \code{matrix} object, each column representing a cell, each row representing a gene.
}
\examples{
sc_data_demo <- demo_sc_data()
}
