% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_st_meta}
\alias{demo_st_meta}
\title{Demo data of st_meta}
\usage{
demo_st_meta()
}
\value{
A data.frame.
}
\description{
Demo data of st_meta
}
\details{
\code{st_meta} used in \code{\link{dec_celltype}} must be a \code{data.frame} object with three columns, namely \code{'spot'}, \code{'x'}, \code{'y'} for spot-based spatial transcriptomics data.
}
\examples{
st_meta_demo <- demo_st_meta()
}
