% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_st_sc_data}
\alias{demo_st_sc_data}
\title{Demo data of single-cell st_data}
\usage{
demo_st_sc_data()
}
\value{
A matrix.
}
\description{
Demo data of single-cell st_data.
}
\details{
\code{st_data} used in \code{\link{dec_celltype}} must be a \code{matrix} object, each column representing a cell, each row representing a gene.
}
\examples{
st_data_demo <- demo_st_sc_data()
}
