% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_st_sc_meta}
\alias{demo_st_sc_meta}
\title{Demo data of st_sc_meta}
\usage{
demo_st_sc_meta()
}
\value{
A data.frame.
}
\description{
Demo data of st_sc_meta
}
\details{
\code{st_sc_meta} used in \code{\link{dec_celltype}} must be a \code{data.frame} object with three columns,  namely \code{'cell'}, \code{'x'}, \code{'y'} for single-cell spatial transcriptomics data.
}
\examples{
st_sc_meta_demo <- demo_st_sc_meta()
}
