% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{DistMatrix-class}
\alias{DistMatrix-class}
\title{Create a 'DistMatrix' object}
\value{
A 'DistMatrix' object.
}
\description{
Create a 'DistMatrix' object.
}
\section{Slots}{

\describe{
\item{\code{pos}}{A matrix containing the x and y coordinates of the cells.}
}}

