% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_lr_path,SpaTalk-method}
\alias{find_lr_path,SpaTalk-method}
\title{Find lrpairs and pathways}
\usage{
\S4method{find_lr_path}{SpaTalk}(object, lrpairs, pathways, max_hop = NULL)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_celltype}}.}

\item{lrpairs}{A data.frame of the system data containing ligand-receptor pairs of \code{'Human'} and \code{'Mouse'} from CellTalkDB.}

\item{pathways}{A data.frame of the system data containing gene-gene interactions and pathways from KEGG and Reactome as well as the information of transcriptional factors.}

\item{max_hop}{Max hop from the receptor to the downstream target transcriptional factor to find for receiving cells. Default is \code{3} for human and \code{4} for mouse.}
}
\value{
SpaTalk object containing the filtered lrpairs and pathways.
}
\description{
Find \code{lrpairs} and \code{pathways} with receptors having downstream targets and transcriptional factors.
}
