% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_lr_path}
\alias{find_lr_path}
\title{Find lrpairs and pathways}
\usage{
find_lr_path(
  object,
  lrpairs,
  pathways,
  max_hop = NULL,
  if_doParallel = T,
  use_n_cores = NULL
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_celltype}}.}

\item{lrpairs}{A data.frame of the system data containing ligand-receptor pairs of \code{'Human'} and \code{'Mouse'} from CellTalkDB.}

\item{pathways}{A data.frame of the system data containing gene-gene interactions and pathways from KEGG and Reactome as well as the information of transcriptional factors.}

\item{max_hop}{Max hop from the receptor to the downstream target transcriptional factor to find for receiving cells. Default is \code{3} for human and \code{4} for mouse.}

\item{if_doParallel}{Use doParallel. Default is TRUE.}

\item{use_n_cores}{Number of CPU cores to use. Default is all cores - 2.}
}
\value{
SpaTalk object containing the filtered lrpairs and pathways.
}
\description{
Find \code{lrpairs} and \code{pathways} with receptors having downstream targets and transcriptional factors.
}
