% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{generate_spot}
\alias{generate_spot}
\title{Generate pseudo spot st_data}
\usage{
generate_spot(st_data, st_meta, x_min, x_res, x_max, y_min, y_res, y_max)
}
\arguments{
\item{st_data}{A data.frame or matrix or dgCMatrix containing counts of spatial transcriptomics, each column representing a cell, each row representing a gene.}

\item{st_meta}{A data.frame containing coordinate of spatial transcriptomics with three columns, \code{'cell'}, \code{'x'}, \code{'y'}, and \code{celltype}.}

\item{x_min}{Min value of x axis.}

\item{x_res}{Resolution of x coordinate.}

\item{x_max}{Max value of x axis.}

\item{y_min}{Min value of y axis.}

\item{y_res}{Resolution of y coordinate.}

\item{y_max}{Max value of y axis.}
}
\value{
A list of spot st_data and st_meta
}
\description{
Generate pseudo spot st_data with single-cell st_data
}
