% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_lr_path}
\alias{get_lr_path}
\title{Get LR and downstream pathways}
\usage{
get_lr_path(
  object,
  celltype_sender,
  celltype_receiver,
  ligand,
  receptor,
  min_gene_num = 5
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_cci}}.}

\item{celltype_sender}{Name of celltype_sender.}

\item{celltype_receiver}{Name of celltype_receiver.}

\item{ligand}{Name of ligand from celltype_sender.}

\item{receptor}{Name of receptor from celltype_receiver.}

\item{min_gene_num}{Min genes number for each pathway.}
}
\value{
A list containing two data.frame. One is LR and downstream pathways, another is the p value of receptor-related pathways with LR-target genes.
}
\description{
Get LR and downstream pathways and get p value of receptor-related pathways with LR-target genes by the Fisher-exact test.
}
