% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{lrpairs}
\alias{lrpairs}
\title{lrpairs}
\format{
An object of class \code{data.frame} with 5427 rows and 3 columns.
}
\source{
\url{http://tcm.zju.edu.cn/celltalkdb/}
}
\usage{
lrpairs
}
\description{
Ligand-receptor pairs of \code{'Human'} and \code{'Mouse'} containing 3398 human and 2033 mouse pairs.
}
\keyword{datasets}
