% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{pathways}
\format{
An object of class \code{data.frame} with 669197 rows and 8 columns.
}
\source{
\url{https://www.genome.jp/kegg/pathway.html}

\url{https://reactome.org/}

\url{http://bioinfo.life.hust.edu.cn/AnimalTFDB/#!/}
}
\usage{
pathways
}
\description{
KEGG pathways and Reactomes of \code{'Human'} and \code{'Mouse'} for intra-cellular genes and transcription factors.
}
\keyword{datasets}
