% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cci_lrpairs,SpaTalk-method}
\alias{plot_cci_lrpairs,SpaTalk-method}
\title{Plot LR pairs}
\usage{
\S4method{plot_cci_lrpairs}{SpaTalk}(
  object,
  celltype_sender,
  celltype_receiver,
  top_lrpairs = 20,
  color = NULL,
  border_color = "black",
  type = NULL,
  fontsize_number = 1,
  number_color = "black"
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_cci}}.}

\item{celltype_sender}{Name of celltype_sender.}

\item{celltype_receiver}{Name of celltype_receiver.}

\item{top_lrpairs}{Number of top lrpairs for plotting. Default is \code{20}.}

\item{color}{Color for the cells in heatmap.}

\item{border_color}{color of cell borders on heatmap, use NA if no border should be drawn.}

\item{type}{Set 'sig' to plot significant LR pairs or set 'number' to plot the number of spatial LR interactions.}

\item{fontsize_number}{fontsize of the numbers displayed in cells.}

\item{number_color}{color of the text.}
}
\description{
Plot LR pairs of celltype_sender and celltype_receiver
}
