% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cci_lrpairs}
\alias{plot_cci_lrpairs}
\title{Plot LR pairs}
\usage{
plot_cci_lrpairs(
  object,
  celltype_sender,
  celltype_receiver,
  top_lrpairs = 20,
  color = NULL,
  border_color = "black",
  type = "sig",
  fontsize_number = 5,
  number_color = "black",
  color_low = NULL,
  color_high = NULL
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_cci}}.}

\item{celltype_sender}{Name of celltype_sender.}

\item{celltype_receiver}{Name of celltype_receiver.}

\item{top_lrpairs}{Number of top lrpairs for plotting. Default is \code{20}.}

\item{color}{Color for the cells in heatmap.}

\item{border_color}{color of cell borders on heatmap, use NA if no border should be drawn.}

\item{type}{Set 'sig' to plot significant LR pairs or set 'number' to plot the number of spatial LR interactions.}

\item{fontsize_number}{fontsize of the numbers displayed in cells.}

\item{number_color}{color of the text.}

\item{color_low}{For 'number' type, define the color for the lowest value.}

\item{color_high}{For 'number' type, define the color for the highest value.}
}
\description{
Heatmap with LR pairs of celltype_sender and celltype_receiver
}
