% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lr_path,SpaTalk-method}
\alias{plot_lr_path,SpaTalk-method}
\title{Plot LR and downstream pathways}
\usage{
\S4method{plot_lr_path}{SpaTalk}(
  object,
  celltype_sender,
  celltype_receiver,
  ligand,
  receptor,
  color = NULL,
  size = 5,
  arrow_length = 0.1
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_cci}}.}

\item{celltype_sender}{Name of celltype_sender.}

\item{celltype_receiver}{Name of celltype_receiver.}

\item{ligand}{Name of ligand from celltype_sender.}

\item{receptor}{Name of receptor from celltype_receiver.}

\item{color}{Color for points Two values.}

\item{size}{Size of points.}

\item{arrow_length}{Arrow length.}
}
\description{
Plot LR and downstream pathways
}
