% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lrpair}
\alias{plot_lrpair}
\title{Plot LR pair}
\usage{
plot_lrpair(
  object,
  celltype_sender,
  celltype_receiver,
  ligand,
  receptor,
  color = NULL,
  size = 1,
  if_plot_density = T,
  if_plot_edge = T,
  arrow_length = 0.05,
  plot_cells = NULL
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_celltype}}.}

\item{celltype_sender}{Name of celltype_sender.}

\item{celltype_receiver}{Name of celltype_receiver.}

\item{ligand}{Name of ligand from celltype_sender.}

\item{receptor}{Name of receptor from celltype_receiver.}

\item{color}{Color for ligand, receptor, and others. Three values.}

\item{size}{Point size. Default is \code{1}.}

\item{if_plot_density}{Whether to plot marginal density plots. Default is \code{TRUE}.}

\item{if_plot_edge}{Whether to plot edge between neighbors. Default is \code{TRUE}.}

\item{arrow_length}{Arrow length.}

\item{plot_cells}{Which cells to plot. Default is all cells. Input a character vector of cell names to plot.}
}
\description{
Point plot with LR pair from celltype_sender to celltype_receiver
}
