% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{plot_lrpair_vln}
\alias{plot_lrpair_vln}
\title{Plot spatial distance of LR pair with vlnplot}
\usage{
plot_lrpair_vln(
  object,
  celltype_sender,
  celltype_receiver,
  ligand,
  receptor,
  vln_color = NULL,
  if_plot_boxplot = T,
  box_width = 0.2
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_celltype}}.}

\item{celltype_sender}{Name of celltype_sender.}

\item{celltype_receiver}{Name of celltype_receiver.}

\item{ligand}{Name of ligand from celltype_sender.}

\item{receptor}{Name of receptor from celltype_receiver.}

\item{vln_color}{Color for violins. Two values.}

\item{if_plot_boxplot}{Whether to plot boxplot. Default is \code{TRUE}.}

\item{box_width}{Box width. Default is \code{0.2}.}
}
\description{
Plot spatial distance of LR pair between expressed senders and receivers and between expressed cell-cell pairs.
}
