% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{plot_path2gene}
\alias{plot_path2gene}
\title{River plot of significantly activated pathways and related downstream genes of receptors.}
\usage{
plot_path2gene(
  object,
  celltype_sender,
  celltype_receiver,
  ligand,
  receptor,
  min_gene_num = 5,
  pvalue = 0.5,
  color = NULL,
  color_flow = "blue"
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_cci}}.}

\item{celltype_sender}{Name of celltype_sender.}

\item{celltype_receiver}{Name of celltype_receiver.}

\item{ligand}{Name of ligand from celltype_sender.}

\item{receptor}{Name of receptor from celltype_receiver.}

\item{min_gene_num}{Min genes number for each pathway.}

\item{pvalue}{P value of the Fisher-exact test.}

\item{color}{Color of pathways and genes. Two values.}

\item{color_flow}{Color of the flow.}
}
\description{
River plot of significantly activated pathways and related downstream genes of receptors.
}
