% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{plot_st_celltype}
\alias{plot_st_celltype}
\title{Plot spatial distribution of a single cell type}
\usage{
plot_st_celltype(
  object,
  celltype,
  size = 1,
  color_celltype = NULL,
  color_others = NULL
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_celltype}}.}

\item{celltype}{Name of cell type in the \code{sc_celltype}.}

\item{size}{Point size. Default is \code{1}.}

\item{color_celltype}{Color for the celltype of interest.}

\item{color_others}{Color for the others.}
}
\description{
Plot spatial distribution of a single predicted cell types for transcriptomics data
}
