% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_st_celltype_all,SpaTalk-method}
\alias{plot_st_celltype_all,SpaTalk-method}
\title{Plot spatial distribution of all cell types}
\usage{
\S4method{plot_st_celltype_all}{SpaTalk}(object, size = 1, color = NULL)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_celltype}}.}

\item{size}{Point size. Default is \code{1}.}

\item{color}{Color for all predicted cell types.}
}
\description{
Plot spatial distribution of all predicted cell types for transcriptomics data
}
