% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_st_celltype_density,SpaTalk-method}
\alias{plot_st_celltype_density,SpaTalk-method}
\title{Plot spatial density of a single cell type}
\usage{
\S4method{plot_st_celltype_density}{SpaTalk}(
  object,
  celltype,
  type,
  if_plot_point = T,
  point_color = NULL,
  point_size = 1,
  color_low = "grey",
  color_mid = NULL,
  color_high = "blue",
  color_midpoint = NULL,
  size = 1
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_celltype}}.}

\item{celltype}{Name of cell type in the \code{sc_celltype}.}

\item{type}{Select 'contour' or 'raster'.}

\item{if_plot_point}{Whether to plot points when type is 'contour'.}

\item{point_color}{Point color.}

\item{point_size}{Point size. Default is \code{1}.}

\item{color_low}{Color for the lowest value.}

\item{color_mid}{Color for the middle value for using \code{scale_color_gradient2}. Default is \code{NULL}.}

\item{color_high}{Color for the highest value.}

\item{color_midpoint}{Value for the middle scale. Default is \code{NULL}.}

\item{size}{Line size when type is 'contour'. Default is \code{1}.}
}
\description{
Plot spatial density of a single predicted cell types for transcriptomics data
}
