% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_st_celltype_percent,SpaTalk-method}
\alias{plot_st_celltype_percent,SpaTalk-method}
\title{Plot spatial distribution of a single cell type percent}
\usage{
\S4method{plot_st_celltype_percent}{SpaTalk}(
  object,
  celltype,
  size = 1,
  color_low = NULL,
  color_mid = NULL,
  color_high = NULL,
  color_midpoint = NULL
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_celltype}}.}

\item{celltype}{Name of cell type in the \code{sc_celltype}.}

\item{size}{Point size. Default is \code{1}.}

\item{color_low}{Color for the lowest value.}

\item{color_mid}{Color for the middle value for using \code{scale_color_gradient2}. Default is \code{NULL}.}

\item{color_high}{Color for the highest value.}

\item{color_midpoint}{Value for the middle scale. Default is \code{NULL}.}
}
\description{
Plot spatial distribution of a single predicted cell type percent for transcriptomics data
}
