% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{plot_st_cor_heatmap}
\alias{plot_st_cor_heatmap}
\title{Plot heatpmap of correlation between marker genes and cell types}
\usage{
plot_st_cor_heatmap(
  object,
  marker_genes,
  celltypes,
  color_low = NULL,
  color_mid = NULL,
  color_high = NULL,
  if_use_newmeta = F,
  scale = "none",
  if_show_top = T,
  top_direction = "row",
  border_color = NA
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_celltype}}.}

\item{marker_genes}{A character containing the known marker genes to plot, provide at least two marker genes of interest.}

\item{celltypes}{A character containing name of cell type in the \code{sc_celltype}. Default is to plot all cell types.}

\item{color_low}{Color for the lowest value.}

\item{color_mid}{Color for the middle value for using \code{scale_color_gradient2}. Default is \code{NULL}.}

\item{color_high}{Color for the highest value.}

\item{if_use_newmeta}{Whether to use newmeta o plot the spatial distribution of gene after \code{\link{dec_celltype}} for spot-based data. Default is \code{FALSE}.}

\item{scale}{Character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. Corresponding values are 'row', 'column' and 'none'.}

\item{if_show_top}{Whether to plot a symbol to the highest value across rows or columns. Default is \code{TRUE}.}

\item{top_direction}{Direction to identify the highest value, select \code{'row'} or \code{'column'}.}

\item{border_color}{Color of the cell border. Default is \code{'NA'}.}
}
\description{
Plot heatpmap of correlation between the expression of marker genes and the predicted score of cell types among all spatial cells or spots.
}
