% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_st_gene}
\alias{plot_st_gene}
\title{Plot spatial distribution of gene}
\usage{
plot_st_gene(
  object,
  gene,
  size = 1,
  color_low = "grey",
  color_mid = NULL,
  color_high = "blue",
  color_midpoint = NULL,
  if_use_newmeta = T,
  celltype = NULL,
  if_plot_others = T
)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_celltype}}.}

\item{gene}{Symbol of gene, e.g., 'AKT1'.}

\item{size}{Point size. Default is \code{1}.}

\item{color_low}{Color for the lowest value.}

\item{color_mid}{Color for the middle value for using \code{scale_color_gradient2}. Default is \code{NULL}.}

\item{color_high}{Color for the highest value.}

\item{color_midpoint}{Value for the middle scale. Default is \code{NULL}.}

\item{if_use_newmeta}{Whether to use newmeta o plot the spatial distribution of gene after \code{\link{dec_celltype}} for spot-based data. Default is \code{TRUE}.}

\item{celltype}{gene in which celltype to plot. Default is \code{NULL}. Set \code{if_use_newmeta} TRUE when using this parameter.}

\item{if_plot_others}{Whether to plot other cells when to use defined \code{celltype}.}
}
\description{
Point plot with spatial distribution of a gene for transcriptomics data
}
\details{
Please set \code{if_use_newmeta} as \code{FALSE} to plot the spatial distribution of gene before \code{\link{dec_celltype}} for spot-based data.
}
