% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{plot_st_pie}
\alias{plot_st_pie}
\title{Plot spatial transcriptomics data}
\usage{
plot_st_pie(object, pie_scale = 1, xy_ratio = 1, color = NULL)
}
\arguments{
\item{object}{SpaTalk object generated from \code{\link{dec_celltype}}.}

\item{pie_scale}{Scale of each pie to plot. Default is \code{1}.}

\item{xy_ratio}{Ratio of y and x coordinates. Default is \code{1}.}

\item{color}{Filled of colors for pie plot, length of \code{color} must be equal to the number of unique cell types in \code{sc_celltype}.}
}
\description{
Plot scatterpie for spatial transcriptomics data
}
