% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_st_pie_generate}
\alias{plot_st_pie_generate}
\title{Plot spatial transcriptomics data}
\usage{
plot_st_pie_generate(st_meta, pie_scale = 1, xy_ratio = 1, color = NULL)
}
\arguments{
\item{st_meta}{st_meta generated from \code{\link{generate_spot}}}

\item{pie_scale}{Scale of each pie to plot. Default is \code{1}.}

\item{xy_ratio}{Ratio of y and x coordinates. Default is \code{1}.}

\item{color}{Filled of colors for pie plot, length of \code{color} must be equal to the number of unique cell types in \code{sc_celltype}.}
}
\description{
Plot scatterpie for spot-based ST data
}
