% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_gene.R
\name{rev_gene}
\alias{rev_gene}
\title{Pre-processing step: revising gene symbols}
\usage{
rev_gene(data = NULL, data_type = NULL, species = NULL, geneinfo = NULL)
}
\arguments{
\item{data}{A data.frame or matrix or dgCMatrix containing count data each column representing a spot or a cell, each row representing a gene; Or a data.frame containing ligand-receptor pairs; Or a data.frame containing gene-gene interactions and pathways from KEGG and Reactome as well as the information of transcriptional factors.}

\item{data_type}{A character to define the type of \code{data}, select \code{'count'} for the data matrix, \code{'lrpairs'} for the data.frame containing lrpairs, \code{'pathways'} for the data.frame containing pathways.}

\item{species}{Species of the data.\code{'Human'} or \code{'Mouse'}.}

\item{geneinfo}{A data.frame of the system data containing gene symbols of \code{'Human'} and \code{'Mouse'} updated on June 30, 2021 for revising count matrix.}
}
\value{
A new matrix or data.frame.
}
\description{
Revise genes according to NCBI Gene symbols updated in June 30, 2021 for count matrix, user-custom lrpairs data.frame, and user-custom pathways data.frame.
}
