% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{set_expected_cell}
\alias{set_expected_cell}
\title{Set the expected cell}
\usage{
set_expected_cell(object, value)
}
\arguments{
\item{object}{SpaTalk object}

\item{value}{Th number of expected cell for each spot, must be equal to the spot number.}
}
\value{
SpaTalk object
}
\description{
Set the expected cell in SpaTalk object
}
