% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show,SpaTalk-method}
\alias{show,SpaTalk-method}
\title{Show SpaTalk object}
\usage{
\S4method{show}{SpaTalk}(object)
}
\arguments{
\item{object}{A SpaTalk object}
}
\value{
Invisible NULL, prints summary information
}
\description{
Show SpaTalk object
}
