% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{SpaTalk}
\alias{SpaTalk}
\alias{SpaTalk-class}
\title{Definition of 'SpaTalk' class}
\description{
An S4 class containing the data, meta, and results of inferred cell type compositions, LR pairs, and pathways.
}
\section{Slots}{

\describe{
\item{\code{data}}{A list containing the raw and normalized data.}

\item{\code{meta}}{A list containing the raw and new meta data.}

\item{\code{para}}{A list containing the parameters.}

\item{\code{coef}}{A matrix containing the results of deconvolution.}

\item{\code{cellpair}}{A list containing the cell-cell pairs based on the spatial distance.}

\item{\code{dist}}{A matrix containing the Euclidean distance among cells.}

\item{\code{lrpair}}{A data frame containing the inferred LR pairs.}

\item{\code{tf}}{A data frame containing the TFs of receptors.}

\item{\code{lr_path}}{A list containing the lrpairs and pathways.}
}}

