% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_batch_cor}
\alias{cpp_batch_cor}
\title{Fast batch correlation}
\usage{
cpp_batch_cor(vec1, mat)
}
\arguments{
\item{vec1}{First vector}

\item{mat}{Matrix where each column is a vector to correlate with vec1}
}
\value{
Correlations
}
\description{
Fast batch correlation
}
