% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_coexp_fast}
\alias{cpp_coexp_fast}
\title{Fast co-expression calculation (C++ vectorized)}
\usage{
cpp_coexp_fast(ligand_expr, receptor_expr)
}
\arguments{
\item{ligand_expr}{Ligand expression matrix (n_genes x n_cells)}

\item{receptor_expr}{Receptor expression matrix (n_genes x n_cells)}
}
\value{
Co-expression ratios
}
\description{
Fast co-expression calculation (C++ vectorized)
}
