% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fast_dist}
\alias{cpp_fast_dist}
\title{Fast Euclidean distance matrix}
\usage{
cpp_fast_dist(x, y)
}
\arguments{
\item{x}{X coordinates}

\item{y}{Y coordinates}
}
\value{
Distance matrix
}
\description{
Fast Euclidean distance matrix
}
