% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fast_sampling}
\alias{cpp_fast_sampling}
\title{Fast cell sampling with early stopping}
\usage{
cpp_fast_sampling(
  spot_ndata,
  sc_ndata_mat,
  cell_indices_by_type,
  spot_celltypes,
  iter_num = 200L,
  tolerance = 0.001,
  seed = 123L
)
}
\arguments{
\item{spot_ndata}{Spot expression vector}

\item{sc_ndata_mat}{SC expression matrix}

\item{cell_indices_by_type}{List of cell indices for each celltype}

\item{spot_celltypes}{Cell types needed for this spot}

\item{iter_num}{Maximum iterations}

\item{tolerance}{Convergence tolerance}

\item{seed}{Random seed}
}
\value{
Best cell combination
}
\description{
Fast cell sampling with early stopping
}
