% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_knn}
\alias{cpp_knn}
\title{Fast K-nearest neighbors}
\usage{
cpp_knn(dist_mat, query_idx, k)
}
\arguments{
\item{dist_mat}{Distance matrix}

\item{query_idx}{Query cell index (0-based)}

\item{k}{Number of neighbors}
}
\value{
Neighbor indices (0-based)
}
\description{
Fast K-nearest neighbors
}
