% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_random_walk}
\alias{cpp_random_walk}
\title{Fast random walk for TF scoring}
\usage{
cpp_random_walk(
  ggi_src,
  ggi_dest,
  receptor_name,
  tf_names,
  n_walks = 10000L,
  max_hop = 10L,
  seed = 123L
)
}
\arguments{
\item{ggi_src}{Source gene names}

\item{ggi_dest}{Destination gene names}

\item{receptor_name}{Starting receptor name}

\item{tf_names}{TF names to score}

\item{n_walks}{Number of walks}

\item{max_hop}{Maximum hops}

\item{seed}{Random seed}
}
\value{
TF scores
}
\description{
Fast random walk for TF scoring
}
